+++ src/libpentobi_mcts/PriorKnowledge.h
@@ -53,7 +53,7 @@ private:
         bool connect;
 
         /** Only used on Classic and Trigon boards. */
-        unsigned dist_to_center;
+        unsigned short dist_to_center;
     };
 
     array<MoveFeatures, Move::range> m_features;
@@ -67,7 +67,7 @@ private:
 
     unsigned m_dist_to_center_max_pieces;
 
-    unsigned m_min_dist_to_center;
+    unsigned short m_min_dist_to_center;
 
     /** Marker for points close to attach points of recent opponent moves.
         0: point is not local, 1: (non-forbidden) oppponent attach point,
@@ -78,7 +78,7 @@ private:
     PointList m_local_points;
 
     /** Distance to center heuristic. */
-    Grid<unsigned> m_dist_to_center;
+    Grid<unsigned short> m_dist_to_center;
 
     void compute_features(const Board& bd, const MoveList& moves,
                           bool check_dist_to_center, bool check_connect);
